move defines to mediapipe utils
